+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "external/raylib"]
-	path = external/raylib
+[submodule "lib/raylib"]
+	path = lib/raylib
 	url = https://github.com/raysan5/raylib
