--- .vscode/settings.json
@@ -1,10 +1,0 @@
-{
-  "files.associations": {
-    "*.ejs": "html",
-    "*.jsx": "javascriptreact",
-    "LICENSE": "plaintext",
-    "*.css": "tailwindcss",
-    "iostream": "cpp",
-    "vector": "cpp"
-  }
-}
