+++ Makefile
@@ -1,37 +1,46 @@
-CXX = clang++
-CXXFLAGS = -Wall -Wextra -O0 -g3 -ggdb3 -std=c++23 -MMD -MP
-INCLUDES = -Iexternal/raylib/src
-LDFLAGS = -Lexternal/raylib/build -lraylib
+NAME		= bombkurdistan
 
-SRCDIR = .
-BUILDDIR = build
-SRC = $(wildcard $(SRCDIR)/*.cpp)
-OBJ = $(patsubst $(SRCDIR)/%.cpp,$(BUILDDIR)/%.o,$(SRC))
-DEP = $(OBJ:.o=.d)
-EXEC = $(BUILDDIR)/bombkurdistan
+INC_DIR		= inc
+LIB_DIR		= lib
+SRC_DIR		= src
+BUILD_DIR	= build
 
-all: raylib $(EXEC)
+RAYLIB_SRC	= $(LIB_DIR)/raylib/src
+RAYLIB		= $(RAYLIB_SRC)/libraylib.a
 
-$(BUILDDIR):
-	mkdir -p $(BUILDDIR)
+CXX			= clang++
+CXXFLAGS	= -Wall -Wextra -O0 -g3 -ggdb3 -std=c++23 -MMD -MP \
+			-I$(INC_DIR) -I$(RAYLIB_SRC)
+LDFLAGS		= -L$(RAYLIB_SRC) -lraylib
+RM			= rm -rf
 
-$(EXEC): $(OBJ) external/raylib/build/raylib/libraylib.a
-	$(CXX) $(CXXFLAGS) $(INCLUDES) -o $@ $^ $(LDFLAGS)
+SRCS		= $(wildcard $(SRC_DIR)/*.cpp)
+OBJS		= $(patsubst $(SRC_DIR)/%.cpp, $(BUILD_DIR)/%.o, $(SRCS))
+DEPS		= $(OBJS:.o=.d)
 
-$(BUILDDIR)/%.o: $(SRCDIR)/%.cpp | $(BUILDDIR)
-	$(CXX) $(CXXFLAGS) $(INCLUDES) -c -o $@ $<
+all: $(NAME)
 
--include $(DEP)
+$(RAYLIB):
+	@$(MAKE) -C $(RAYLIB_SRC) PLATFORM=PLATFORM_DESKTOP
 
-external/raylib/build/raylib/libraylib.a:
-	mkdir -p external/raylib/build
-	cd external/raylib/build && \
-	cmake .. -DBUILD_SHARED_LIBS=OFF -DCMAKE_BUILD_TYPE=Release && \
-	make
+$(BUILD_DIR):
+	@mkdir -p $@
 
-raylib: external/raylib/build/raylib/libraylib.a
+$(BUILD_DIR)/%.o: $(SRC_DIR)/%.cpp | $(BUILD_DIR)
+	@$(CXX) $(CXXFLAGS) -c $< -o $@
+
+$(NAME): $(OBJS) $(RAYLIB)
+	@$(CXX) $(CXXFLAGS) $(OBJS) $(LDFLAGS) -o $@
 
 clean:
-	rm -rf $(BUILDDIR) external/raylib/build
+	@$(RM) $(BUILD_DIR)
 
-.PHONY: all clean raylib
+fclean: clean
+	@$(MAKE) -C $(RAYLIB_SRC) clean
+	@$(RM) $(NAME)
+
+re: fclean all
+
+-include $(DEPS)
+
+.PHONY: all clean fclean re
